refactor: Extract useInitSetup and useAppDataSetup
